⚡️ :: CheckBox State 파일 분류
